+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -447,6 +447,7 @@ sub build_isf_public_interface_contract {
             't/1278-isf-enum-member-transaction-aggregate-params.t',
             't/1279-isf-enum-member-inline-drive-values.t',
             't/1280-isf-enum-member-inline-drive-expression-values.t',
+            't/1281-isf-enum-member-library-use-params.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',